+++ lib/speculations/parser/state/in.rb
@@ -1,14 +1,14 @@
 module Speculations
   class Parser
     module State
-      module Bef extend self
+      module In extend self
 
-        def parse line, _lnb, node
+        def parse line, _lnb, node, ctxt
           case
           when State.eoblock_match(line)
             [:out, node.parent]
           else
-            [:bef, node.add_line(line)]
+            [:in, node.add_line(line)]
           end
         end
       end