+++ lib/speculations/parser/state/out.rb
@@ -3,71 +3,26 @@ module Speculations
     module State
       module Out extend self
 
-        def parse line, lnb, node
-          if node.alternate_syntax
-            _parse_alternate line, lnb, node
+        def parse line, lnb, node, _ctxt
+          case
+          when match = State.context_match(line)
+            make_new_context(lnb: lnb, node: node, match: match)
+          when match = State.maybe_example(line)
+            [:candidate, node, match[:title]]
+          when match = State.maybe_include(line)
+            [:candidate, node, :inc]
           else
-            _parse_classical line, lnb, node
+            [:out, node]
           end
         end
 
         private
 
-        def _parse_alternate line, lnb, node
-          case
-          when name = State.context_match(line)
-            node = node.parent if node.parent
-            new_node = node.add_child(name: name, lnb: lnb)
-            [:out, new_node.reset_context]
-          when State.ws_match(line)
-            [:out, node]
-          when name = State.potential_name(line)
-            node.set_name(name[1])
-            node.set_given(false)
-            [:out, node]
-          when name = State.then_match(line)
-            node.set_name(name[1])
-            node.set_given(false)
-            [:out, node]
-          when State.given_match(line)
-            node.set_name(nil)
-            [:out, node.set_given(true)]
-          when State.ruby_match(line)
-            if node.potential_name
-              node = node.add_example(lnb: lnb, line: line)
-              [:exa, node]
-            elsif node.given?
-              node = node.add_include(lnb: lnb)
-              [:inc, node]
-            else
-              [:out, node]
-            end
-          else
-            [:out, node.reset_context]
-          end
-        end
-
-        def _parse_classical line, lnb, node
-          case
-          when name = State.context_match(line)
-            node = node.parent if node.parent
-            new_node = node.add_child(name: name, lnb: lnb)
-            [:out, new_node]
-          when State.before_match(line)
-            node = node.set_setup(lnb: lnb)
-            [:bef, node]
-          when State.example_match(line)
-            node = node.add_example(lnb: lnb, line: line)
-            [:exa, node]
-          when State.include_match(line)
-            node = node.add_include(lnb: lnb)
-            [:inc, node]
-          when name = State.potential_name(line)
-            node.set_name(name[1])
-            [:out, node]
-          else
-            [:out, node]
-          end
+        def make_new_context(lnb:, match:, node:)
+          level = match[:level].size
+          new_parent = node.parent_of_level(level.pred)
+          node = new_parent.new_context(title: match[:title], lnb: lnb, level: level)
+          [:out, node]
         end
 
       end