+++ spec/speculate_about/alternate_syntax_spec.rb
@@ -1,3 +1,0 @@
-RSpec.describe "spec/fixtures/ALTERNATE_SYNTAX.md" do
-  speculate_about description, alternate_syntax: true
-end