+++ spec/speculate_about/readme_spec.rb
@@ -1,3 +1,0 @@
-RSpec.describe "README.md" do
-  speculate_about description
-end