+++ spec/speculate_about/speculate_about_spec.rb
@@ -1,6 +1,0 @@
-RSpec.describe "Speculations v1" do
-
-  speculate_about "./V1.md"
-  speculate_about "../fixtures/*.md"
-
-end